Fix type mismatch

+++ web/server.go
@@ -43,11 +43,11 @@ func Start() {
 		var nextCursor int64
 		var previousCursor int64
 		if len(completedJobs) == int(options.Limit) {
-			nextCursor = completedJobs[len(completedJobs)-1].ID
+			nextCursor = int64(completedJobs[len(completedJobs)-1].ID)
 			if cursor != "" {
 				ii, err := strconv.ParseInt(cursor, 10, 0)
 				OrPanic(err)
-				previousCursor = ii + options.Limit
+				previousCursor = ii + int64(options.Limit)
 			}
 		}
